requirements.txt
pandas
selenium
beautifulsoup4
gspread
google-auth-oauthlib
requests
python-dotenv
